wndr3700v4

+++ my_profile.sh
@@ -4,8 +4,8 @@
 function my_install_x86_64
 {
 	sudo apt-get install -y lib32z1 lib32ncurses5 lib32bz2-1.0 lib32stdc++6 \
-		build-essential ccache bison flex automake intltool libncurses5-dev \
-		unzip git bash bash-completion vim
+		build-essential ccache bison flex automake intltool libncurses5-dev libssl-dev \
+		unzip git subversion mercurial gawk bash bash-completion vim
 }
 
 # 解包zip
@@ -43,6 +43,26 @@ function my_unpack_tgz
 			wget -O ${DL_FILE} ${DL_URL};
 		fi;
 		tar -zxf ${DL_FILE};
+		rm -rf ${DL_FILE};
+	fi;
+	popd;
+}
+
+# 解包tar.bz2
+# UNPACK_DIR 必须是完整目录路径
+function my_unpack_tar_bz2
+{
+	DL_URL=$2
+	UNPACK_DIR=$1
+	DL_FILE=${DL_URL##*/};
+	
+	mkdir -p ${UNPACK_DIR%/*};
+	pushd ${UNPACK_DIR%/*};
+	if [ ! -d "${UNPACK_DIR}" ]; then
+		if [ ! -f "${DL_FILE}" ]; then
+			wget -O ${DL_FILE} ${DL_URL};
+		fi;
+		tar -jxf ${DL_FILE};
 		rm -rf ${DL_FILE};
 	fi;
 	popd;
